Compartmentalized
Moved search bar into a separate header component
