+++ src/Grunnkart/Grunnkart.js
@@ -8,6 +8,7 @@ import { SettingsContext } from "../SettingsContext";
 import språk from "../språk";
 import VenstreVinduContainer from "../VenstreVinduContainer";
 import bakgrunnskarttema from "./bakgrunnskarttema";
+import TopBar from "../TopBar/TopBar";
 
 const styles = {
   rot: {
@@ -209,34 +210,28 @@ class Grunnkart extends React.Component {
           const transparent = isAtRoot(history.location);
           return (
             <>
-              <div
-                className={classNames(
-                  classes.rot,
-                  transparent && classes.transparent
-                )}
-              >
-                <VenstreVinduContainer
-                  aktiveLag={this.state.aktiveLag}
-                  mapBounds={this.state.actualBounds}
-                  onMouseEnter={this.handleMouseEnter}
-                  onMouseLeave={this.handleMouseLeave}
-                  onFitBounds={this.handleFitBounds}
-                  erAktivert={erAktivert}
-                  opplystKode={this.state.opplystKode}
-                  onToggleLayer={() => {
-                    this.handleToggleLayer();
-                    if (!context.visAktiveLag) context.onToggleAktiveLag();
-                  }}
-                  onRemoveSelectedLayer={this.handleRemoveSelectedLayer}
-                  meta={this.state.meta}
-                  searchFor={this.state.searchFor}
-                  onClearSearchFor={this.handleClearSearchFor}
-                  onUpdateLayerProp={this.handleUpdateLayerProp}
-                  onUpdateMetaProp={this.handleUpdateMetaProp}
-                  visAktiveLag={context.visAktiveLag}
-                  onToggleAktiveLag={context.onToggleAktiveLag}
-                />
-              </div>
+              <TopBar />
+              <VenstreVinduContainer
+                aktiveLag={this.state.aktiveLag}
+                mapBounds={this.state.actualBounds}
+                onMouseEnter={this.handleMouseEnter}
+                onMouseLeave={this.handleMouseLeave}
+                onFitBounds={this.handleFitBounds}
+                erAktivert={erAktivert}
+                opplystKode={this.state.opplystKode}
+                onToggleLayer={() => {
+                  this.handleToggleLayer();
+                  if (!context.visAktiveLag) context.onToggleAktiveLag();
+                }}
+                onRemoveSelectedLayer={this.handleRemoveSelectedLayer}
+                meta={this.state.meta}
+                searchFor={this.state.searchFor}
+                onClearSearchFor={this.handleClearSearchFor}
+                onUpdateLayerProp={this.handleUpdateLayerProp}
+                onUpdateMetaProp={this.handleUpdateMetaProp}
+                visAktiveLag={context.visAktiveLag}
+                onToggleAktiveLag={context.onToggleAktiveLag}
+              />
               <Kart
                 bounds={this.state.fitBounds}
                 latitude={65.4}
