+++ src/LookupControl/SearchField/SearchField.js
